backend/requirements.txt
Flask==3.1.2
flask_cors==6.0.2
python-dotenv
requests
docker